文件模板:1.根据文件名排序；2.修改命名规则；3.每页展示25条。 高原体检：1.过期天数要可以更新；2.修改是否过期。 图纸归档状态：1.上传文件数大于等于指定归档文件数即算归档成功。

+++ src/main/java/com/ruoyi/project/system/archive/service/impl/ProjectArchiveServiceImpl.java
@@ -1,41 +1,44 @@
 package com.ruoyi.project.system.archive.service.impl;
 
-import java.io.File;
-import java.util.List;
-
 import com.ruoyi.common.exception.ServiceException;
 import com.ruoyi.common.utils.DateUtils;
+import com.ruoyi.common.utils.text.Convert;
+import com.ruoyi.project.system.archifile.domain.ProjectArchifile;
+import com.ruoyi.project.system.archifile.mapper.ProjectArchifileMapper;
+import com.ruoyi.project.system.archive.domain.ProjectArchive;
+import com.ruoyi.project.system.archive.mapper.ProjectArchiveMapper;
+import com.ruoyi.project.system.archive.service.IProjectArchiveService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import com.ruoyi.project.system.archive.mapper.ProjectArchiveMapper;
-import com.ruoyi.project.system.archive.domain.ProjectArchive;
-import com.ruoyi.project.system.archive.service.IProjectArchiveService;
-import com.ruoyi.common.utils.text.Convert;
+
+import java.io.File;
+import java.util.List;
 
 import static com.ruoyi.common.utils.security.ShiroUtils.getLoginName;
 import static com.ruoyi.common.utils.security.ShiroUtils.getSysUser;
 
 /**
  * 项目归档Service业务层处理
- * 
+ *
  * @author ruoyi
  * @date 2023-09-18
  */
 @Service
-public class ProjectArchiveServiceImpl implements IProjectArchiveService 
-{
+public class ProjectArchiveServiceImpl implements IProjectArchiveService {
     @Autowired
     private ProjectArchiveMapper projectArchiveMapper;
 
+    @Autowired
+    private ProjectArchifileMapper projectArchifileMapper;
+
     /**
      * 查询项目归档
-     * 
+     *
      * @param archiveId 项目归档主键
      * @return 项目归档
      */
     @Override
-    public ProjectArchive selectProjectArchiveByArchiveId(Long archiveId)
-    {
+    public ProjectArchive selectProjectArchiveByArchiveId(Long archiveId) {
         return projectArchiveMapper.selectProjectArchiveByArchiveId(archiveId);
     }
 
@@ -104,6 +107,10 @@ public class ProjectArchiveServiceImpl implements IProjectArchiveService
     @Override
     public int deleteProjectArchiveByArchiveIds(String archiveIds)
     {
+        List<ProjectArchifile> files = projectArchifileMapper.selectProjectArchifileListByArid(archiveIds);
+        if (!files.isEmpty()) {
+            throw new ServiceException("文件夹内存在项目文件，请手动删除文件再删除子表记录！");
+        }
         return projectArchiveMapper.deleteProjectArchiveByArchiveIds(Convert.toStrArray(archiveIds));
     }
